Added an updatecommand.

+++ Source/Frontend/Managecommands.cpp
@@ -21,7 +21,7 @@ void Printusage(std::vector<Frontend::Argument_t> Arguments)
     for (const auto &Item : *Commands)
     {
         Infoprint(va("%-15s  %-.63s", Item.first.c_str(), Item.second.Description.c_str()));
-        for (int i = 1; i < 4; ++i)
+        for (size_t i = 1; i < 4; ++i)
         {
             if (Item.second.Description.size() > 80 * i)
             {
